% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrond_10.R
\name{arrond_10}
\alias{arrond_10}
\title{arrondi à la dizaine.}
\usage{
arrond_10(x)
}
\arguments{
\item{x}{vecteur numerique}
}
\value{
vecteur numerique
}
\description{
arrondi à la dizaine.
}
\examples{
arrond_10(56)
}
