% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrond_100.R
\name{arrond_100}
\alias{arrond_100}
\title{arrondi a la centaine.}
\usage{
arrond_100(x)
}
\arguments{
\item{x}{vecteur numerique}
}
\value{
vecteur numerique
}
\description{
arrondi a la centaine.
}
\examples{
arrond_100(1198L)
}
