% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cree_carte_terrains.R
\name{cree_carte_terrains}
\alias{cree_carte_terrains}
\title{Dessine une carte}
\usage{
cree_carte_terrains(
  data = toto,
  x = "terr_prixmoy_m2",
  y = "Prix moyen des terrains en euros"
)
}
\arguments{
\item{data}{dataframe}

\item{x}{caractere axe des x}

\item{y}{caractere axe des y}
}
\value{
plot
}
\description{
Dessine une carte
}
