% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn01_choix_annee_etude.R
\name{fn01_choix_annee_etude}
\alias{fn01_choix_annee_etude}
\title{Choix annee etude}
\usage{
fn01_choix_annee_etude(annee = 2019)
}
\arguments{
\item{annee}{un vecteur numerique de longueur 4 caracteres}
}
\description{
La fonction enregistre dans une liste l'annee d'etude et calcule l'annee
precedente au format texte
}
\examples{
fn01_choix_annee_etude(2018)

}
