% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn01_cree_repertoires.R
\name{fn01_cree_repertoires}
\alias{fn01_cree_repertoires}
\title{cree les repertoires dans le repertoire de travail.}
\usage{
fn01_cree_repertoires(
  new_rep = c("1_scripts", "2_data", "3_tables", "4_resultats", "5_publication")
)
}
\arguments{
\item{new_rep}{vecteur de caracteres avec le nom des repertoires a creer}
}
\value{
nothing cree les repertoires
}
\description{
cree les repertoires dans le repertoire de travail.
}
