% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn02_choix_annee_etude.R
\name{fn02_choix_annee_etude}
\alias{fn02_choix_annee_etude}
\title{Choix annee etude}
\usage{
fn02_choix_annee_etude(x = 2019)
}
\arguments{
\item{x}{un vecteur numerique de longueur 4 caracteres}
}
\description{
La fonction enregistre dans une liste l'annee d'etude et calcule l'annee
precedente au format texte
}
\examples{
fn02_choix_annee_etude(2018)

}
