% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn03_lit_tab_beyond.R
\name{fn03_import_tab_beyond}
\alias{fn03_import_tab_beyond}
\title{Lit les deux tableaux issus de beyond}
\usage{
fn03_import_tab_beyond(x = "2_data")
}
\arguments{
\item{x}{nom du repertoire ou se trouvent les 2 fichiers csv (format caractere)}
}
\value{
une liste de tables
}
\description{
Lit le repertoire, cherche les deux fichiers csv extraits manuellement
par l'utilisateur sur le site du sdes au format beyond puis les transforme
NB pourrait etre ajoute à la requete sous geokit
}
