% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn04_import_tableaux_geokit.R
\name{fn04_import_tableaux_geokit}
\alias{fn04_import_tableaux_geokit}
\title{Lit les tableaux geokit}
\usage{
fn04_import_tableaux_geokit(x = "2_data")
}
\arguments{
\item{x}{nom du repertoire ou se trouvent les 2 fichiers csv (character)}
}
\value{
une liste de tables
}
\description{
Lit le repertoire, cherche le fichiers xlsx issus de geokit
ouvre le fichier, lit les onglets
}
