% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn05_transforme_tab_gk_filtre_frregdep.R
\name{fn05_transforme_tab_gk_filtre_frregdep}
\alias{fn05_transforme_tab_gk_filtre_frregdep}
\title{Normalise les tableaux geokit terrains pour les differents territoires
meme nombre de colonnes et memes champs}
\usage{
fn05_transforme_tab_gk_filtre_frregdep(
  fich_frregdep_filtre = c("filtre_fr", "filtre_reg", "filtre_dep")
)
}
\arguments{
\item{fich_frregdep_filtre}{vecteur de caracteres}
}
\value{

}
\description{
Normalise les tableaux geokit terrains pour les differents territoires
meme nombre de colonnes et memes champs
}
