% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn06_met_en_forme_liste.R
\name{fn06_met_en_forme_liste}
\alias{fn06_met_en_forme_liste}
\title{Met en forme les tableaux}
\usage{
fn06_met_en_forme_liste(x, y)
}
\arguments{
\item{x}{charactere}

\item{y}{charactere}
}
\value{

}
\description{
Met en forme les tableaux
}
