% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn08_lit_fich_filtre_var.R
\name{fn08_lit_fich_filtre_var}
\alias{fn08_lit_fich_filtre_var}
\title{Terrains : lit les tableaux geokit par modalite csp etc...}
\usage{
fn08_lit_fich_filtre_var(x = "filtre_csp")
}
\arguments{
\item{x}{un caractere specifiant la modalite}
}
\value{

}
\description{
Terrains : lit les tableaux geokit par modalite csp etc...
}
