% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn09_extrait_tab_filtre_modalites.R
\name{fn09_extrait_tab_filtre_modalites}
\alias{fn09_extrait_tab_filtre_modalites}
\title{Terrains : cree une liste de tableau par modalites csp etc}
\usage{
fn09_extrait_tab_filtre_modalites(
  fich_filtre = c("filtre_csp", "filtre_rp", "filtre_age", "filtre_viabilisation")
)
}
\arguments{
\item{fich_filtre}{un vecteur de caractere avec toutes les modalites}
}
\value{
liste
}
\description{
Terrains : cree une liste de tableau par modalites csp etc
}
