% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn10_tab_filtre_verifie_secstat.R
\name{fn10_tab_filtre_verifie_secstat}
\alias{fn10_tab_filtre_verifie_secstat}
\title{Terrains : verifie si le secret statistique s'applique}
\usage{
fn10_tab_filtre_verifie_secstat(data = tli_reg_indics_filtre)
}
\arguments{
\item{data}{une liste de tables}
}
\value{

}
\description{
Terrains : verifie si le secret statistique s'applique
}
