% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn11_tab_filtre_applique_secstat.R
\name{fn11_tab_filtre_applique_secstat}
\alias{fn11_tab_filtre_applique_secstat}
\title{Terrains : applique le secret statistique et masque les informations sensibles}
\usage{
fn11_tab_filtre_applique_secstat(data = tli_reg_indics_filtre)
}
\arguments{
\item{data}{une liste de tables}
}
\value{

}
\description{
Terrains : applique le secret statistique et masque les informations sensibles
}
