% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn12_transforme_tab_gk_ssfiltre_frregdep.R
\name{fn12_transforme_tab_gk_ssfiltre_frregdep}
\alias{fn12_transforme_tab_gk_ssfiltre_frregdep}
\title{Maisons : normalise les tableaux geokit maisons pour les differents territoires.
toto.}
\usage{
fn12_transforme_tab_gk_ssfiltre_frregdep(
  fich_frregdep_ssfiltre = c("ssfiltre_fr", "ssfiltre_reg", "ssfiltre_dep")
)
}
\arguments{
\item{fich_frregdep_ssfiltre}{un vecteur de caracteres}
}
\value{

}
\description{
Maisons : normalise les tableaux geokit maisons pour les differents territoires.
toto.
}
