% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn13_cree_doc_frcordep_ssfiltre.R
\name{fn13_cree_doc_frcordep_ssfiltre}
\alias{fn13_cree_doc_frcordep_ssfiltre}
\title{Maisons : modifie les tableaux maisons pour la plaquette}
\usage{
fn13_cree_doc_frcordep_ssfiltre(data = tli_regdep_ssfiltre)
}
\arguments{
\item{data}{liste de tableaux}
}
\value{

}
\description{
Maisons : modifie les tableaux maisons pour la plaquette
}
