% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn14_lit_fich_ssfiltre_var.R
\name{fn14_lit_fich_ssfiltre_var}
\alias{fn14_lit_fich_ssfiltre_var}
\title{Maisons : lit les tableaux geokit par modalite csp etc...}
\usage{
fn14_lit_fich_ssfiltre_var(x = "ssfiltre_csp")
}
\arguments{
\item{x}{un caractere specifiant la modalite}
}
\value{

}
\description{
Maisons : lit les tableaux geokit par modalite csp etc...
}
