% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn15_extrait_tab_ssfiltre_modalites.R
\name{fn15_extrait_tab_ssfiltre_modalites}
\alias{fn15_extrait_tab_ssfiltre_modalites}
\title{Maisons : cree une liste de tableau par modalites csp moe chauffage etc}
\usage{
fn15_extrait_tab_ssfiltre_modalites(
  fich_ssfiltre = c("ssfiltre_csp", "ssfiltre_rp", "ssfiltre_age", "ssfiltre_finition",
    "ssfiltre_moe", "ssfiltre_chauffage")
)
}
\arguments{
\item{fich_ssfiltre}{un vecteur de caractere avec toutes les modalites}
}
\value{
liste
}
\description{
Maisons : cree une liste de tableau par modalites csp moe chauffage etc
}
