% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn16_tab_ssfiltre_verifie_secstat.R
\name{fn16_tab_ssfiltre_verifie_secstat}
\alias{fn16_tab_ssfiltre_verifie_secstat}
\title{Maisons : verifie si le secret statistique s'applique}
\usage{
fn16_tab_ssfiltre_verifie_secstat(data = tli_reg_indics_ssfiltre)
}
\arguments{
\item{data}{une liste de tables}
}
\value{

}
\description{
Maisons : verifie si le secret statistique s'applique
}
