% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn17_tab_ssfiltre_applique_secstat.R
\name{fn17_tab_ssfiltre_applique_secstat}
\alias{fn17_tab_ssfiltre_applique_secstat}
\title{Maisons : applique le secret statistique et masque les donnees sensibles}
\usage{
fn17_tab_ssfiltre_applique_secstat(data = tli_reg_indics_ssfiltre)
}
\arguments{
\item{data}{une liste de tables}
}
\value{

}
\description{
Maisons : applique le secret statistique et masque les donnees sensibles
}
