% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn18_calcul_evolution_maisons_terrains.R
\name{fn18_calcul_evolution_maisons_terrains}
\alias{fn18_calcul_evolution_maisons_terrains}
\title{Mons et terrains : calcule les evolutions sur un an et depuis 2010}
\usage{
fn18_calcul_evolution_maisons_terrains(
  data = list(doc_frcordep_filtre = doc_frcordep_filtre, doc_frcordep_ssfiltre =
    doc_frcordep_ssfiltre)
)
}
\arguments{
\item{data}{une data frame}
}
\value{

}
\description{
Mons et terrains : calcule les evolutions sur un an et depuis 2010
}
