% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn19_calcul_evolution_maisons_terrains_mod.R
\name{fn19_calcul_evolution_maisons_terrains_mod}
\alias{fn19_calcul_evolution_maisons_terrains_mod}
\title{Terrains et maisons : ajoute la comparaison avec France entière}
\usage{
fn19_calcul_evolution_maisons_terrains_mod(
  data = tli_frcordepcor_filtre_ssfiltre$evol_an
)
}
\arguments{
\item{data}{une dataframe}
}
\value{

}
\description{
Terrains et maisons : ajoute la comparaison avec France entière
}
