% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn20_mef_tableaux_plaquette.R
\name{fn20_mef_tableaux_plaquette}
\alias{fn20_mef_tableaux_plaquette}
\title{Plaquette  : Réalisation de la plaquette}
\usage{
fn20_mef_tableaux_plaquette(
  data = list(doc_frcordep_filtre = doc_frcordep_filtre, tab2_terrains = tab2_terrains,
    doc_frcordep_ssfiltre = doc_frcordep_ssfiltre, tab4_maisons = tab4_maisons)
)
}
\arguments{
\item{data}{liste de dataframe}
}
\value{

}
\description{
Enregistre le contenu des tableaux avec les indicateurs calcules
}
