% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn21_enregistre_tableaux.R
\name{fn21_enregistre_tableaux}
\alias{fn21_enregistre_tableaux}
\title{Plaquette : enregistre les tableaux de calculs complémentaires}
\usage{
fn21_enregistre_tableaux(x = ls_dates$annee_etude)
}
\arguments{
\item{x}{annee etude en caracteres}
}
\value{

}
\description{
Plaquette : enregistre les tableaux de calculs complémentaires
}
