% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn30_cartos_terrains_indics.R
\name{fn30_carto_terrains_indics}
\alias{fn30_carto_terrains_indics}
\title{Terrains : cartographie}
\usage{
fn30_carto_terrains_indics(data = tli_regdep_filtre[["filtre_reg"]])
}
\arguments{
\item{data}{dataframe}
}
\value{

}
\description{
F1 Préparation de la cartographie pour les terrains des régions
métropolitaines filtre.
calcul pour l'annee d'étudiée des étiquettes de la carte représentée
sous forme de quartile
}
