% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn31_carte_maisons.R
\name{fn31_carte_maisons}
\alias{fn31_carte_maisons}
\title{Maisons : cartographie}
\usage{
fn31_carte_maisons(data = tli_regdep_ssfiltre[["ssfiltre_reg"]])
}
\arguments{
\item{data}{dataframe}
}
\value{
nothing
}
\description{
Maisons : cartographie
}
