% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn32_graphiques_terrains.R
\name{fn32_graphiques_terrains}
\alias{fn32_graphiques_terrains}
\title{Terrains : les graphiques}
\usage{
fn32_graphiques_terrains(data = doc_frcordep_filtre)
}
\arguments{
\item{data}{dataframe}
}
\value{

}
\description{
Terrains : les graphiques
}
