% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn33_graphiques_maisons.R
\name{fn33_graphiques_maisons}
\alias{fn33_graphiques_maisons}
\title{Maisons : les graphiques}
\usage{
fn33_graphiques_maisons(data = doc_frcordep_ssfiltre)
}
\arguments{
\item{data}{dataframe}
}
\value{

}
\description{
Maisons : les graphiques
}
