% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn33b_cowplot_maison_terrain.R
\name{fn33b_cowplot_maison_terrain}
\alias{fn33b_cowplot_maison_terrain}
\title{Terrains et maisons : on colle deux graphiques}
\usage{
fn33b_cowplot_maison_terrain(
  x = ls_cowplot$plot_terr_prixm2,
  y = ls_cowplot$plot_mai_prixm2
)
}
\arguments{
\item{x}{un plot}

\item{y}{un plot}
}
\value{

}
\description{
Terrains et maisons : on colle deux graphiques
}
