% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn34_autres_graphiques_terrains.R
\name{fn34_autres_graphiques_terrains}
\alias{fn34_autres_graphiques_terrains}
\title{Terrains : autres graphiques non utilises dans la plaquette}
\usage{
fn34_autres_graphiques_terrains(data = tab2_terrains)
}
\arguments{
\item{data}{dataframe}
}
\value{

}
\description{
Terrains : autres graphiques non utilises dans la plaquette
}
