% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn35_autres_graphiques_maisons.R
\name{fn35_autres_graphiques_maisons}
\alias{fn35_autres_graphiques_maisons}
\title{Maisons : autres graphiques non utilises dans la plaquette}
\usage{
fn35_autres_graphiques_maisons(data = tab4_maisons)
}
\arguments{
\item{data}{dataframe}
}
\value{

}
\description{
Maisons : autres graphiques non utilises dans la plaquette
}
