% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn40_scr_lit_nouveau_texte.R
\name{fn40_scr_lit_nouveau_texte}
\alias{fn40_scr_lit_nouveau_texte}
\title{Scribus : texte, lit le texte de remplacement}
\usage{
fn40_scr_lit_nouveau_texte(x = ls_dates$annee_etude)
}
\arguments{
\item{x}{annee etude en caratere}
}
\value{
liste
}
\description{
Scribus : texte, lit le texte de remplacement
}
