% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn41_lit_le_modele.R
\name{fn41_lit_le_modele}
\alias{fn41_lit_le_modele}
\title{Scribus : texte, , lit la plaquette modele}
\usage{
fn41_lit_le_modele(x = ls_dates$annee_etude)
}
\arguments{
\item{x}{date annee etude}
}
\value{

}
\description{
Scribus : texte, , lit la plaquette modele
}
