% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn42_remplace_txt_paragraphe.R
\name{fn42_remplace_txt_paragraphe}
\alias{fn42_remplace_txt_paragraphe}
\title{Scibus : texte, remplace le texte du paragraphe}
\usage{
fn42_remplace_txt_paragraphe(data = ls_modele$t_parametres)
}
\arguments{
\item{data}{une tibble avec les colonnes num_objet, par, lgn}
}
\value{

}
\description{
Scibus : texte, remplace le texte du paragraphe
}
