% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn42b_remplace_txt_titre.R
\name{fn42b_remplace_txt_titre}
\alias{fn42b_remplace_txt_titre}
\title{Scibus : texte, remplace le texte du titre}
\usage{
fn42b_remplace_txt_titre(x = ls_dates$annee_etude)
}
\arguments{
\item{x}{la date de l'annee d'etude}
}
\value{

}
\description{
Scibus : texte, remplace le texte du titre
}
