% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn43_lit_contenu_tableaux.R
\name{fn43_lit_contenu_tableaux}
\alias{fn43_lit_contenu_tableaux}
\title{Scribus : tableaux, lit le contenu des tableaux du modele}
\usage{
fn43_lit_contenu_tableaux(x = 2018)
}
\arguments{
\item{x}{vecteur}
}
\value{
liste
}
\description{
Scribus : tableaux, lit le contenu des tableaux du modele
}
