% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn44_tableaux_cellule_a_remplacer.R
\name{fn44_tableaux_cellule_a_remplacer}
\alias{fn44_tableaux_cellule_a_remplacer}
\title{Scribus : tableaux, tableau des cellules a remplacer}
\usage{
fn44_tableaux_cellule_a_remplacer(x = resultats[3:6])
}
\arguments{
\item{x}{liste}
}
\value{
df
}
\description{
Scribus : tableaux, tableau des cellules a remplacer
}
