% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn45_tableaux_remplace_corps_tableau.R
\name{fn45_tableaux_remplace_corps_tableau}
\alias{fn45_tableaux_remplace_corps_tableau}
\title{scribus, Tableaux, ajoute les valeurs issues des calculs dans les tableaux}
\usage{
fn45_tableaux_remplace_corps_tableau(x = resultats[3:6])
}
\arguments{
\item{x}{liste}
}
\value{
rien
}
\description{
scribus, Tableaux, ajoute les valeurs issues des calculs dans les tableaux
}
