% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn46_tableaux_met_en_forme_champs.R
\name{fn46_tableaux_met_en_forme_champs}
\alias{fn46_tableaux_met_en_forme_champs}
\title{Scribus, tableaux, mise en forme de la première ligne du tableau}
\usage{
fn46_tableaux_met_en_forme_champs(x = ls_tableaux0$t_tableaux_cell)
}
\arguments{
\item{x}{liste}
}
\value{
rien
}
\description{
Scribus, tableaux, mise en forme de la première ligne du tableau
}
