% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn47_tableaux_met_en_forme_ssindicateurs.R
\name{fn47_tableaux_met_en_forme_ssindicateurs}
\alias{fn47_tableaux_met_en_forme_ssindicateurs}
\title{Scribus, tableaux, mise en forme des sous-indicateurs}
\usage{
fn47_tableaux_met_en_forme_ssindicateurs(x = toto)
}
\arguments{
\item{x}{df}
}
\value{
rien
}
\description{
Enregistre aussi toutes les modifications en sortie
}
