% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn90_classement.R
\name{fn90_classement}
\alias{fn90_classement}
\title{classement}
\usage{
fn90_classement(x)
}
\arguments{
\item{x}{un vecteur numerique}
}
\value{
un vecteur numerique
}
\description{
classement
}
\examples{
fn90_classement(c(5,4,6,8,1,3))
}
