% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn91_terrains_calcul_5indics.R
\name{fn91_terrains_calcul_5indics}
\alias{fn91_terrains_calcul_5indics}
\title{Terrains calcule des 5 indicateurs de la plaquette}
\usage{
fn91_terrains_calcul_5indics(data)
}
\arguments{
\item{data}{tableau}
}
\value{
tableau modifié
}
\description{
Terrains calcule des 5 indicateurs de la plaquette
}
