% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn92_terrains_calcul_part.R
\name{fn92_terrains_calcul_part}
\alias{fn92_terrains_calcul_part}
\title{Terrains calcul de la part de chaque territoire
(en nombre de terrains etudies)}
\usage{
fn92_terrains_calcul_part(data)
}
\arguments{
\item{data}{tableau}
}
\value{
tableau modifié
}
\description{
Terrains calcul de la part de chaque territoire
(en nombre de terrains etudies)
}
