% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn93_maisons_calcul_3indics.R
\name{fn93_maisons_calcul_3indics}
\alias{fn93_maisons_calcul_3indics}
\title{Maisons calcule des 3 indicateurs de la plaquette}
\usage{
fn93_maisons_calcul_3indics(data)
}
\arguments{
\item{data}{tableau}
}
\value{
tableau modifié
}
\description{
Maisons calcule des 3 indicateurs de la plaquette
}
