% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn94_maisons_calcul_part.R
\name{fn94_maisons_calcul_part}
\alias{fn94_maisons_calcul_part}
\title{Maisons calcul de la part de chaque territoire
(en nombre de terrains etudies)}
\usage{
fn94_maisons_calcul_part(data)
}
\arguments{
\item{data}{tableau}
}
\value{
tableau modifié
}
\description{
Maisons calcul de la part de chaque territoire
(en nombre de terrains etudies)
}
