% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn95_lit_fich_depreg_filtre.R
\name{fn95_lit_fich_depreg_filtre}
\alias{fn95_lit_fich_depreg_filtre}
\title{Terrains lit les onglets du fichier excel}
\usage{
fn95_lit_fich_depreg_filtre(x)
}
\arguments{
\item{x}{un vecteur avec les references des onglets}
}
\value{
une liste
}
\description{
Terrains lit les onglets du fichier excel
}
