% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn96_diff_an.R
\name{fn96_diff_an}
\alias{fn96_diff_an}
\title{Difference entre deux annees}
\usage{
fn96_diff_an(x, annee)
}
\arguments{
\item{x}{une table avec une colonne annee et une colonne pour x}

\item{annee}{une colonne annee}
}
\value{

}
\description{
Difference entre deux annees
}
