% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rc_po_no_td_numcell.R
\name{fn_rc_po_no_td_numcell}
\alias{fn_rc_po_no_td_numcell}
\title{Scribus interroge le fichier sla pour extraire tous les nodes d'une cellule
d'un tableau.}
\usage{
fn_rc_po_no_td_numcell(num_objet = 6, num_cell = 2)
}
\arguments{
\item{num_objet}{id du tableau}

\item{num_cell}{id de la cellule}
}
\value{
nodes de la cellule
}
\description{
Scribus interroge le fichier sla pour extraire tous les nodes d'une cellule
d'un tableau.
}
