% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dec.R
\name{format_dec}
\alias{format_dec}
\title{Formate les nombres decimaux au format texte}
\usage{
format_dec(x)
}
\arguments{
\item{x}{vecteur}
}
\value{

}
\description{
Formate les nombres decimaux au format texte
}
\examples{
format_dec(102.65)
}
