% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pourcent.R
\name{format_pourcent}
\alias{format_pourcent}
\title{Pourcentage au format texte}
\usage{
format_pourcent(x)
}
\arguments{
\item{x}{vecteur}
}
\value{

}
\description{
Pourcentage au format texte
}
\examples{
format_pourcent(0.05)
}
