% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_carto.R
\docType{data}
\name{ls_carto}
\alias{ls_carto}
\title{les couches des regions et des pays europeens}
\format{
a list of two sf object and a table

\describe{
\item{map_eu}{couche des pays}
\item{map_reg}{couche des regions}
\item{map_reg_pt}{table des coordonnees des centroides des regions}
}
}
\source{
\url{IGN}
}
\usage{
ls_carto
}
\description{
Donnees IGN
}
\keyword{gis}
