% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_calculs_metadonnees.R
\docType{data}
\name{tab_calculs_metadonnees}
\alias{tab_calculs_metadonnees}
\title{Table explicitant les codes indicateurs}
\format{
table

\describe{
\item{champ_indic}{code indicateur}
\item{description}{description}
\item{unité}{unite}
\item{remarques}{remarques}
}
}
\source{
ODDC
}
\usage{
data("tab_calculs_metadonnees")
}
\description{
La table ne sert pas dans le code mais est ensuite copiee dans le tableau excell de sortie
}
\keyword{metadata}
