% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_plaquette.R
\docType{data}
\name{txt_plaquette}
\alias{txt_plaquette}
\title{Le fichier xml issu de scribus au format texte}
\format{
texte
}
\usage{
data("txt_plaquette")
}
\description{
Modele de document qui sera rempli par les autres fonctions du pkg
}
\author{
G.Winterstein
}
\keyword{scribus}
